save loaded data to minimize the time for 2nd load

+++ lib/make_view.rb
@@ -17,27 +17,39 @@ class MakeView
     @html_data
   end
   def read_log(dir_name)
-    File.open(@log_root + dir_name + "/route.json", "r") do |f|
-      @json_data[:route_data] = JSON.load(f)
-      @json_data[:route_data] = @json_data[:route_data].map{|v| ((v[0] < @config[:secure_area][0][0] && v[1] > @config[:secure_area][0][1]) && (v[0] > @config[:secure_area][1][0] && v[1] < @config[:secure_area][1][1]))? nil : v}.compact
-      @json_data[:start] = @json_data[:route_data][0]
-      @json_data[:end] = @json_data[:route_data][-1]
+    if(File.exist?(@log_root + dir_name + "/contents.json"))
+      File.open(@log_root + dir_name + "/contents.json", "r") do |f|
+        @json_data = JSON.load(f,nil, {:symbolize_names => true})
+      end
+    else 
+      File.open(@log_root + dir_name + "/route.json", "r") do |f|
+        @json_data[:route_data] = JSON.load(f)
+        @json_data[:route_data] = @json_data[:route_data].map{|v| ((v[0] < @config[:secure_area][0][0] && v[1] > @config[:secure_area][0][1]) && (v[0] > @config[:secure_area][1][0] && v[1] < @config[:secure_area][1][1]))? nil : v}.compact
+        @json_data[:start] = @json_data[:route_data][0]
+        @json_data[:end] = @json_data[:route_data][-1]
+      end
+      File.open(@log_root + dir_name + "/config.json", "r") do |f|
+        @json_data.merge!(JSON.load(f, nil, {:symbolize_names => true}))
+      end
+      mid_lat = (@json_data[:route_data].map{|v| v[0]}.max + @json_data[:route_data].map{|v| v[0]}.min)/2
+      mid_long = (@json_data[:route_data].map{|v| v[1]}.max + @json_data[:route_data].map{|v| v[1]}.min)/2
+      @json_data[:midpoint] = [mid_lat, mid_long]
+      File.open(@log_root + dir_name + "/contents.json", "w") do |f|
+        JSON.dump(@json_data,f)
+      end
     end
-    File.open(@log_root + dir_name + "/config.json", "r") do |f|
-      @json_data.merge!(JSON.load(f, nil, {:symbolize_names => true}))
-    end
-    mid_lat = (@json_data[:route_data].map{|v| v[0]}.max + @json_data[:route_data].map{|v| v[0]}.min)/2
-    mid_long = (@json_data[:route_data].map{|v| v[1]}.max + @json_data[:route_data].map{|v| v[1]}.min)/2
-    @json_data[:midpoint] = [mid_lat, mid_long]
   end
   def get_json
     return @json_data
   end
   def make_html
     @html_data = @erb.result(binding)
-    File.open("tmp.html","w") do |f|
-      f.puts(@html_data)
-    end
+#    File.open("tmp.html","w") do |f|
+#      f.puts(@html_data)
+#    end
     true
+  end
+  def get_html
+    return @html_data
   end
 end
